backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-docx
pdfminer.six
pandas
qdrant-client
numpy
python-dotenv
requests
